Reorganized files for Flipper App Catalog structure
