--- app_types.h
@@ -1,9 +1,0 @@
-#pragma once
-
-// Forward declarations only
-typedef struct AppState AppState;
-typedef struct UartContext UartContext;
-typedef struct StorageContext StorageContext;
-typedef struct SettingsUIContext SettingsUIContext;
-typedef struct ConfirmationView ConfirmationView;
-typedef struct UartStorageContext UartStorageContext;
