--- ghost_esp_icons.h
@@ -1,8 +1,0 @@
-#pragma once
-#include <gui/icon.h>
-
-extern const Icon I_Wifi_icon;
-extern const Icon I_BLE_icon; 
-extern const Icon I_GPS_icon;
-extern const Icon I_Cog;
-extern const Icon I_ButtonDown_7x4;
